+++ incuna_test_utils/compat.py
@@ -18,9 +18,19 @@ def wipe_id_fields_on_django_lt_17(fields):
     return fields
 
 
-class Python2CountEqualMixin(object):
-    """Add assertCountEqual to python 2 TestCase"""
+class Python2AssertMixin(object):
+    """
+    Add python 3 asserts to python 2 TestCase
+
+    Asserts added:
+    * assertCountEqual
+    * assertRegex
+    """
+
     def __init__(self, *args, **kwargs):
-        super(Python2CountEqualMixin, self).__init__(*args, **kwargs)
+        super(Python2AssertMixin, self).__init__(*args, **kwargs)
         if not hasattr(self, 'assertCountEqual'):
             self.assertCountEqual = self.assertItemsEqual
+
+        if not hasattr(self, 'assertRegex'):
+            self.assertRegex = self.assertRegexpMatches